% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear the Cache}
\usage{
clear_cache()
}
\value{
Nothing. Deletes cached files.
}
\description{
Removes all cached returns from the PokéAPI, which are stored as RDS files
at the path resolved by \code{tools::R_user_dir("trapinch", "cache")}.
}
\examples{
\dontrun{clear_cache()}
}
