% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_berry_firmness}
\alias{get_berry_firmness}
\title{Get Data from PokeAPI: Berry Firmnesses}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
get_berry_firmness(resource)
}
\arguments{
\item{resource}{Character or numeric. The name or ID of the item to return
from the named endpoint.}
}
\value{
A (usually-) nested list.
}
\description{
Get Data from PokeAPI: Berry Firmnesses
}
\details{
See \href{https://pokeapi.co/docs/v2#berries}{the API documentation}.
}
\examples{
\dontrun{get_berry_firmness("very-soft")}
}
