% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_evolution_trigger}
\alias{get_evolution_trigger}
\title{Get Data from PokeAPI: Evolution Triggers}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
get_evolution_trigger(resource)
}
\arguments{
\item{resource}{Character or numeric. The name or ID of the item to return
from the named endpoint.}
}
\value{
A nested list.
}
\description{
Get Data from PokeAPI: Evolution Triggers
}
\details{
See \href{https://pokeapi.co/docs/v2#berries}{the API documentation}.
}
\examples{
\dontrun{get_evolution_trigger("level-up")}
}
