% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_item_attribute}
\alias{get_item_attribute}
\title{Get Data from PokeAPI: Item Attributes}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
get_item_attribute(resource)
}
\arguments{
\item{resource}{Character or numeric. The name or ID of the item to return
from the named endpoint.}
}
\value{
A nested list.
}
\description{
Get Data from PokeAPI: Item Attributes
}
\details{
See \href{https://pokeapi.co/docs/v2#berries}{the API documentation}.
}
\examples{
\dontrun{get_item_attribute("countable")}
}
