% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_item_category}
\alias{get_item_category}
\title{Get Data from PokeAPI: Item Categories}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
get_item_category(resource)
}
\arguments{
\item{resource}{Character or numeric. The name or ID of the item to return
from the named endpoint.}
}
\value{
A nested list.
}
\description{
Get Data from PokeAPI: Item Categories
}
\details{
See \href{https://pokeapi.co/docs/v2#berries}{the API documentation}.
}
\examples{
\dontrun{get_item_category("stat-boosts")}
}
