% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_location_area}
\alias{get_location_area}
\title{Get Data from PokeAPI: Location Areas}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
get_location_area(resource)
}
\arguments{
\item{resource}{Character or numeric. The name or ID of the item to return
from the named endpoint.}
}
\value{
A (usually-) nested list.
}
\description{
Get Data from PokeAPI: Location Areas
}
\details{
See \href{https://pokeapi.co/docs/v2#berries}{the API documentation}.
}
\examples{
\dontrun{get_location_area("canalave-city-area")}
}
