% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_move_battle_style}
\alias{get_move_battle_style}
\title{Get Data from PokeAPI: Move Battle Styles}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
get_move_battle_style(resource)
}
\arguments{
\item{resource}{Character or numeric. The name or ID of the item to return
from the named endpoint.}
}
\value{
A (usually-) nested list.
}
\description{
Get Data from PokeAPI: Move Battle Styles
}
\details{
See \href{https://pokeapi.co/docs/v2#berries}{the API documentation}.
}
\examples{
\dontrun{get_move_battle_style("attack")}
}
