% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_move_learn_method}
\alias{get_move_learn_method}
\title{Get Data from PokeAPI: Move Learn Methods}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
get_move_learn_method(resource)
}
\arguments{
\item{resource}{Character or numeric. The name or ID of the item to return
from the named endpoint.}
}
\value{
A (usually-) nested list.
}
\description{
Get Data from PokeAPI: Move Learn Methods
}
\details{
See \href{https://pokeapi.co/docs/v2#berries}{the API documentation}.
}
\examples{
\dontrun{get_move_learn_method("level-up")}
}
