% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_pokeapi}
\alias{get_pokeapi}
\title{Get Data from the PokéAPI Service}
\source{
API by PokéAPI \url{https://pokeapi.co/}. Data from The Pokémon Company.
}
\usage{
get_pokeapi(endpoint, resource = NULL, ext = NULL, verbose = FALSE)
}
\arguments{
\item{endpoint}{Character. The name of the endpoint to query. This includes
categories such as \code{"pokemon"}, \code{"locations"} or
\code{"berries"}. See details.}

\item{resource}{Character or numeric. The name or ID of the resource to
return from the named endpoint. Most 'get' functions can accept either,
but some only accept an ID. Defaults to \code{NULL}, in which case all
possible resources are returned for the provided \code{endpoint}. See
details.}

\item{ext}{Character. Further extension to the provided endpoint and
resource. Defaults to \code{NULL}. Only used for the 'pokemon' endpoint
to find Pokémon encounters.}

\item{verbose}{Logical. Show extra API-related output when request is
performed? Defaults to \code{FALSE}.}
}
\value{
A list. The exact contents depend on the endpoint and resource being
requested
}
\description{
Get Data from the PokéAPI Service
}
\details{
Note that the 'trapinch' package uses version 2 of the API. See
\href{https://pokeapi.co/docs/v2}{the PokéAPI documentation} for more
information and a full list of endpoints and resources. You can also view
the built-in \code{\link{resource_lookups}} data for this information.
Results are cached.
}
\examples{
\dontrun{get_pokeapi("pokemon", "bulbasaur")}
}
