% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_pokemon_shape}
\alias{get_pokemon_shape}
\title{Get Data from PokeAPI: Pokemon Shapes}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
get_pokemon_shape(resource)
}
\arguments{
\item{resource}{Character or numeric. The name or ID of the item to return
from the named endpoint.}
}
\value{
A (usually-) nested list.
}
\description{
Get Data from PokeAPI: Pokemon Shapes
}
\details{
See \href{https://pokeapi.co/docs/v2#berries}{the API documentation}.
}
\examples{
\dontrun{get_pokemon_shape("ball")}
}
