% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_super_contest_effect}
\alias{get_super_contest_effect}
\title{Get Data from PokeAPI: Super Contest Effects}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
get_super_contest_effect(resource)
}
\arguments{
\item{resource}{Character or numeric. The name or ID of the item to return
from the named endpoint.}
}
\value{
A nested list.
}
\description{
Get Data from PokeAPI: Super Contest Effects
}
\details{
See \href{https://pokeapi.co/docs/v2#berries}{the API documentation}.
}
\examples{
\dontrun{get_super_contest_effect(1)}
}
