% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_version_group}
\alias{get_version_group}
\title{Get Data from PokeAPI: Version Groups}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
get_version_group(resource)
}
\arguments{
\item{resource}{Character or numeric. The name or ID of the item to return
from the named endpoint.}
}
\value{
A nested list.
}
\description{
Get Data from PokeAPI: Version Groups
}
\details{
See \href{https://pokeapi.co/docs/v2#berries}{the API documentation}.
}
\examples{
\dontrun{get_version_group("red-blue")}
}
