% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resource_lookups}
\alias{resource_lookups}
\title{All Resources and Endpoints}
\format{
A list of dataframes of length 48, which is one per endpoint
available from PokéAPI. Each dataframe has columns \code{id} (integer),
\code{name} (character) (possibly) and \code{url} (character), and a
number of rows equal to the total number of resources available for the
given endpoint.
}
\source{
API by PokéAPI \url{https://pokeapi.co/}. Data from The Pokémon Company.
}
\usage{
resource_lookups
}
\description{
IDs, names and URLs for all the resources for each endpoint that can be
requested from the PokéAPI.
}
\keyword{datasets}
