% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapinch-package.R
\docType{package}
\name{trapinch-package}
\alias{trapinch}
\alias{trapinch-package}
\title{trapinch: Get Data from the PokéAPI Service}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Query the RESTful PokéAPI <https://pokeapi.co/> to return information about the Pokemon game franchise, including moves, berries, items, locations, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/trapinch}
  \item Report bugs at \url{https://github.com/matt-dray/trapinch/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
