% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_pokeapi}
\alias{get_pokeapi}
\title{Get Data from the PokéAPI Service}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
get_pokeapi(endpoint, resource, ext = NULL)
}
\arguments{
\item{endpoint}{Character. The name of the endpoint to query, such as
\code{"berries"}, \code{"locations"} or \code{"pokemon"}.}

\item{resource}{Character or numeric. The name or ID of the item to return
from the named endpoint.}

\item{ext}{Character. Only needed for Pokémon encounters.}
}
\value{
A nested list.
}
\description{
Get Data from the PokéAPI Service
}
\details{
See \href{https://pokeapi.co/docs/v2#berries}{the API documentation}.
}
\examples{
\dontrun{get_pokeapi("pokemon", "bulbasaur")}
}
