% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_super_contest_effect}
\alias{get_super_contest_effect}
\title{Get Data from PokeAPI: Super Contest Effects}
\source{
API by PokéAPI \url{https://pokeapi.co/}. Data from The Pokémon Company.
}
\usage{
get_super_contest_effect(resource)
}
\arguments{
\item{resource}{Character or numeric. The name or ID of the resource to
return from the named endpoint. Most 'get' functions can accept either,
but some only accept an ID.. See details.}
}
\value{
A (usually-) nested list. The contents depend on the endpoint and
resource being called.
}
\description{
Get Data from PokeAPI: Super Contest Effects
}
\details{
See \href{https://pokeapi.co/docs/v2}{the PokéAPI documentation}.for
more information and a full list of endpoints and resources. You can also
view the built-in \code{\link{resource_lookups}} data for this
information. Note that the 'trapinch' package uses version 2 of the API.
}
\examples{
\dontrun{get_super_contest_effect(1)}
}
