% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resource_lookups}
\alias{resource_lookups}
\title{All Endpoint Resources}
\format{
A list of dataframes of length 48, which is the number of endpoints
available from PokéAPI. Each dataframe has columns \code{id}, \code{name}
(possibly) and \code{url}, and a number of rows equal to the total number
of resources for the given endpoint.
}
\source{
PokéAPI \url{https://pokeapi.co/}.
}
\usage{
resource_lookups
}
\description{
IDs, names and URLs for all the resources for each endpoint that can be
requested from the PokéAPI.
}
\keyword{datasets}
